+++ frontend/src/components/form/styles.ts
@@ -1,0 +1,26 @@
+import {StyleSheet} from 'react-native';
+
+export const formStyles = StyleSheet.create({
+  container: {
+    gap: 30,
+  },
+  formContainer: {
+    gap: 15,
+  },
+  bottomContainer: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+    gap: 10,
+  },
+  bottomText: {
+    fontSize: 14,
+    fontWeight: '600',
+    color: '#999',
+  },
+  bottomLink: {
+    fontSize: 14,
+    textDecorationLine: 'underline',
+    fontWeight: '600',
+    color: '#5d8fc5',
+  },
+});
